backend/services/auth.py
import os
import firebase_admin
from firebase_admin import credentials, auth
from fastapi import Depends, HTTPException
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials

security = HTTPBearer()

def initialize_firebase():
    """Start the Firebase Admin Engine. Gracefully degrade if key is missing locally."""
    creds_path = os.getenv("FIREBASE_CREDENTIALS_PATH", "firebase_key.json")
    
    if os.path.exists(creds_path):
        try:
            # Check if app is already initialized (FastAPI reload safeguard)
            firebase_admin.get_app()
        except ValueError:
            print(f"Initializing Firebase Admin with key from {creds_path}")
            cred = credentials.Certificate(creds_path)
            firebase_admin.initialize_app(cred)
    else:
        print(f"Warning: Firebase Admin Key '{creds_path}' not found! Production auth checks will fail.")


def verify_firebase_token(creds: HTTPAuthorizationCredentials = Depends(security)) -> dict:
    """Verifies a generic Bearer string against Firebase servers."""
    token = creds.credentials
    
    # Fast path for automated tests 
    if token == "mock-token-for-tests":
        return {"uid": "mock-test-uid-123", "email": "test@example.com"}

    try:
        # Request Google server token validation
        decoded_token = auth.verify_id_token(token)
        return decoded_token
    except Exception as e:
        raise HTTPException(
            status_code=401,
            detail=f"Invalid or expired authentication credentials: {e}",
            headers={"WWW-Authenticate": "Bearer"},
        )
