backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
google-generativeai==0.8.0
google-cloud-vision==3.7.2
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
Pillow==10.4.0
pytest==8.3.2
